+++ client/src/components/posts.css
@@ -1,0 +1,4 @@
+.titles, .buttons {
+    text-align: left !important;
+    width: 100%;
+}
