+++ client/src/components/posts.jsx
@@ -1,6 +1,8 @@
 import { useEffect, useState } from 'react';
 import ShowPost from './showPost';
 import Form from './form';
+import './posts.css';
+import { Link } from 'react-router-dom';
 
 function Posts() {
     const [posts, setPosts] = useState([]);
@@ -13,20 +15,6 @@ function Posts() {
                 setPosts(posts);
                 console.log('Posts fetched...', posts);
             });
-    }
-
-    const handlePostRequest = (data) => {
-        console.log("Inside the POST, ", data);
-        fetch("http://localhost:8080/posts", {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(data)
-        })
-        .then((response) => response.json())
-        .then((data) => {
-            console.log("In the final stretch, ", data);
-            setPosts([...posts, data]);
-        })
     }
 
     const showArticle = (post) => {
@@ -42,7 +30,11 @@ function Posts() {
         <div>
             <div data-testid='post-list'>
                 {posts.map((post) => (
-                    <div key={post.id}><button onClick={() => showArticle(post)}>{post.title}</button></div>
+                    <div className='titles' key={post.id}>
+                        <Link to={`/article/${post.id}`}>
+                            <button className='buttons' onClick={() => showArticle(post)}>{post.title}</button>
+                        </Link>
+                    </div>
                 ))}
                 {selectedPost && <ShowPost post={selectedPost} data-testid='selected-post' />}
             </div>
